+++ app/configwindow.cpp
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2026 KIO Rclone contributors
+ * SPDX-FileCopyrightText: 2026 Gabriel Maizo González <maizogabriel@gmail.com>
  *
  * SPDX-License-Identifier: GPL-2.0-or-later
  */